ficket-face/requirements.txt
albucore==0.0.20
albumentations==1.4.21
annotated-types==0.7.0
blinker==1.8.2
boto3==1.35.58
botocore==1.35.58
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
coloredlogs==15.0.1
contourpy==1.3.0
cryptography==43.0.3
cycler==0.12.1
Cython==3.0.11
dnspython==2.7.0
easydict==1.13
eval_type_backport==0.2.0
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
flatbuffers==24.3.25
fonttools==4.54.1
greenlet==3.1.1
humanfriendly==10.0
idna==3.10
ifaddr==0.2.0
imageio==2.36.0
insightface==0.7.3
itsdangerous==2.2.0
javaobj-py3==0.4.4
Jinja2==3.1.4
jmespath==1.0.1
joblib==1.4.2
kiwisolver==1.4.7
lazy_loader==0.4
MarkupSafe==3.0.2
matplotlib==3.9.2
mpmath==1.3.0
mysql-connector==2.2.9
mysql-connector-python==9.1.0
networkx==3.4.2
numpy==2.1.3
onnx==1.17.0
onnxruntime==1.20.0
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
packaging==24.2
pika==1.3.2
pillow==11.0.0
prettytable==3.12.0
protobuf==5.28.3
py_eureka_client==0.11.12
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pycryptodome==3.21.0
pycryptodomex==3.21.0
pydantic==2.9.2
pydantic_core==2.23.4
pyjks==20.0.0
PyMySQL==1.1.1
pyparsing==3.2.0
pyreadline3==3.5.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.2
requests==2.32.3
s3transfer==0.10.3
scikit-image==0.24.0
scikit-learn==1.5.2
scipy==1.14.1
setuptools==75.3.0
simsimd==6.0.3
six==1.16.0
SQLAlchemy==2.0.36
stringzilla==3.10.9
sympy==1.13.3
threadpoolctl==3.5.0
tifffile==2024.9.20
tqdm==4.67.0
twofish==0.3.0
typing_extensions==4.12.2
urllib3==2.2.3
wcwidth==0.2.13
Werkzeug==3.1.2
